+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add documentation example for the allocator
+
+### Fixed
+- Fix extension ordering
 
 ## [0.6.0+140] - 2020-05-04
 ### Added