+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add documentation example for the allocator
 - Add `DefaultCoreLoader` alias
+- Better builder support for void pointers
 
 ### Changed
 - Rename `erupt-examples` to `erupt_examples`